requirements.txt
Babel==2.3.4
click==6.6
debtcollector==1.5.0
Flask==0.11.1
Flask-MySQLdb==0.2.0
funcsigs==1.0.2
futures==3.0.5
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
keystoneauth1==2.9.0
MarkupSafe==0.23
monotonic==1.1
msgpack-python==0.4.7
mysqlclient==1.3.7
netaddr==0.7.18
netifaces==0.10.4
oslo.config==3.12.0
oslo.i18n==3.7.0
oslo.serialization==2.10.0
oslo.utils==3.15.0
pbr==1.10.0
positional==1.1.1
pyDes==2.0.1
python-keystoneclient==3.2.0
python-swiftclient==3.0.0
pytz==2016.4
requests==2.10.0
rfc3986==0.3.1
six==1.10.0
stevedore==1.15.0
Werkzeug==0.11.10
wrapt==1.10.8
